+++ src/components/Drop/Drop.Positioner.jsx
@@ -12,8 +12,12 @@ import getValidProps from '@helpscout/react-utils/dist/getValidProps'
 import ReactDOM from 'react-dom'
 import EventListener from '../EventListener'
 import classNames from 'classnames'
-import { applyStylesToNode, isNodeElement } from '../../utilities/node'
-import { getViewportPosition, getDirections } from './Drop.utils'
+import { isNodeElement } from '../../utilities/node'
+import {
+  getViewportPosition,
+  getDirections,
+  applyStylesToNode,
+} from './Drop.utils'
 
 import { DropContentUI, DropUI } from './Drop.css'
