--- src/utilities/__tests__/node/getScrollParent.test.js
@@ -1,8 +1,0 @@
-import { getScrollParent } from '../../node'
-
-test('Returns null for invalid arguments', () => {
-  expect(getScrollParent()).toBe(null)
-  expect(getScrollParent(true)).toBe(null)
-  expect(getScrollParent('div')).toBe(null)
-  expect(getScrollParent({})).toBe(null)
-})
