--- src/utilities/__tests__/node/getViewportHeight.test.js
@@ -1,6 +1,0 @@
-import { getViewportHeight } from '../../node'
-
-test('Returns height as a number', () => {
-  expect(getViewportHeight()).toBeTruthy()
-  expect(typeof getViewportHeight()).toBe('number')
-})
