--- src/utilities/__tests__/node/getViewportWidth.test.js
@@ -1,6 +1,0 @@
-import { getViewportWidth } from '../../node'
-
-test('Returns width as a number', () => {
-  expect(getViewportWidth()).toBeTruthy()
-  expect(typeof getViewportWidth()).toBe('number')
-})
