--- src/utilities/__tests__/node/hasContentOverflowY.test.js
@@ -1,9 +1,0 @@
-import { hasContentOverflowY } from '../../node'
-
-test('Returns false for invalid elements', () => {
-  expect(hasContentOverflowY()).not.toBeTruthy()
-  expect(hasContentOverflowY(true)).not.toBeTruthy()
-  expect(hasContentOverflowY('div')).not.toBeTruthy()
-  expect(hasContentOverflowY(window)).not.toBeTruthy()
-  expect(hasContentOverflowY({})).not.toBeTruthy()
-})
